+++ src/validate_async.cc
@@ -3,27 +3,25 @@
 #include "v8_wrapper.h"
 #include "protagonist.h"
 #include "snowcrash.h"
-#include "drafter.h"
 
 using std::string;
 using namespace v8;
 using namespace protagonist;
 
-// Async Parse
-void AsyncParse(uv_work_t* request);
+// Async Validate
+void AsyncValidate(uv_work_t* request);
 
-// Async Parse Handler
-void AsyncParseAfter(uv_work_t* request);
+// Async Validate Handler
+void AsyncValidateAfter(uv_work_t* request);
 
 // Threadpooling libuv baton
-struct ParseBaton {
+struct ValidateBaton {
 
     // Callback
     Nan::Persistent<Function> callback;
 
     // Input
     snowcrash::BlueprintParserOptions options;
-    drafter::ASTType astType;
     mdp::ByteBuffer sourceData;
 
     // Output
@@ -31,29 +29,29 @@ struct ParseBaton {
     sos::Object result;
 };
 
-NAN_METHOD(protagonist::Parse) {
+NAN_METHOD(protagonist::Validate) {
     Nan::HandleScope scope;
 
     // Check arguments
     if (info.Length() != 2 && info.Length() != 3) {
-        Nan::ThrowTypeError("wrong number of arguments, `parse(string, options, callback)` expected");
+        Nan::ThrowTypeError("wrong number of arguments, `validate(string, options, callback)` expected");
         return;
     }
 
     if (!info[0]->IsString()) {
-        Nan::ThrowTypeError("wrong argument - string expected, `parse(string, options, callback)`");
+        Nan::ThrowTypeError("wrong argument - string expected, `validate(string, options, callback)`");
         return;
     }
 
     if ((info.Length() == 2 && !info[1]->IsFunction()) ||
         (info.Length() == 3 && !info[2]->IsFunction())) {
 
-        Nan::ThrowTypeError("wrong argument - callback expected, `parse(string, options, callback)`");
+        Nan::ThrowTypeError("wrong argument - callback expected, `validate(string, options, callback)`");
         return;
     }
 
     if (info.Length() == 3 && !info[1]->IsObject()) {
-        Nan::ThrowTypeError("wrong argument - object expected, `parse(string, options, callback)`");
+        Nan::ThrowTypeError("wrong argument - object expected, `validate(string, options, callback)`");
         return;
     }
 
@@ -62,10 +60,9 @@ NAN_METHOD(protagonist::Parse) {
 
     // Prepare options
     snowcrash::BlueprintParserOptions options = 0;
-    drafter::ASTType astType = drafter::RefractASTType;
 
     if (info.Length() == 3) {
-        OptionsResult *optionsResult = ParseOptionsObject(Handle<Object>::Cast(info[1]), false);
+        OptionsResult *optionsResult = ParseOptionsObject(Handle<Object>::Cast(info[1]), true);
 
         if (optionsResult->error != NULL) {
             Nan::ThrowTypeError(optionsResult->error);
@@ -73,7 +70,6 @@ NAN_METHOD(protagonist::Parse) {
         }
 
         options = optionsResult->options;
-        astType = optionsResult->astType;
         free(optionsResult);
     }
 
@@ -81,9 +77,8 @@ NAN_METHOD(protagonist::Parse) {
     Local<Function> callback = (info.Length() == 3) ? Local<Function>::Cast(info[2]) : Local<Function>::Cast(info[1]);
 
     // Prepare threadpool baton
-    ParseBaton* baton = ::new ParseBaton();
+    ValidateBaton* baton = ::new ValidateBaton();
     baton->options = options;
-    baton->astType = astType;
     baton->sourceData = *sourceData;
     baton->callback.Reset(callback);
 
@@ -94,36 +89,34 @@ NAN_METHOD(protagonist::Parse) {
     // Schedule the work request
     int status = uv_queue_work(uv_default_loop(),
                                 request,
-                                AsyncParse,
-                                (uv_after_work_cb)AsyncParseAfter);
+                                AsyncValidate,
+                                (uv_after_work_cb)AsyncValidateAfter);
 
     assert(status == 0);
     return;
 }
 
-void AsyncParse(uv_work_t* request) {
-    ParseBaton* baton = static_cast<ParseBaton*>(request->data);
+void AsyncValidate(uv_work_t* request) {
+    ValidateBaton* baton = static_cast<ValidateBaton*>(request->data);
 
-    // Parse the source data
-    snowcrash::parse(baton->sourceData, baton->options | snowcrash::ExportSourcemapOption, baton->parseResult);
-    baton->result = Result::WrapResult(baton->parseResult, baton->options, baton->astType);
+    baton->result = Validate::Do(baton->sourceData, baton->options);
 }
 
-void AsyncParseAfter(uv_work_t* request) {
+void AsyncValidateAfter(uv_work_t* request) {
     Nan::HandleScope scope;
-    ParseBaton* baton = static_cast<ParseBaton*>(request->data);
+    ValidateBaton* baton = static_cast<ValidateBaton*>(request->data);
 
     // Prepare report
     const unsigned argc = 2;
     Local<Value> argv[argc];
 
-    argv[1] = v8_wrap(baton->result)->ToObject();
+    // Error Object
+    argv[0] = Nan::Null();
 
-    // Error Object (FIXME: Don't have this once we remove AST)
-    if (baton->parseResult.report.error.code == snowcrash::Error::OK)
-        argv[0] = Nan::Null();
+    if (baton->result.empty())
+        argv[1] = Nan::Null();
     else
-        argv[0] = SourceAnnotation::WrapSourceAnnotation(baton->parseResult.report.error);
+        argv[1] = v8_wrap(baton->result)->ToObject();
 
     TryCatch try_catch;
     Local<Function> callback = Nan::New<Function>(baton->callback);