+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Protagonist Changelog
 
+## Master
+
+### Enhancements
+
+* Added `validate` and `validateSync` to just return the warnings and errors
+  after parsing a blueprint.
+
 ## 1.5.2
 
 This update now uses Drafter 3.1.3. Please see [Drafter